fix: set dtype of timestep to float in PriorTransformer

+++ src/optimum/rbln/diffusers/models/transformers/prior_transformer.py
@@ -162,7 +162,7 @@ class RBLNPriorTransformer(RBLNModel):
     ):
         return self.runtime.forward(
             hidden_states.contiguous(),
-            timestep.long(),
+            timestep.float(),
             proj_embedding,
             encoder_hidden_states,
             attention_mask.float(),
